Fix: 게시물 삭제 기능 트랜잭션 추가

+++ src/domains/post/postController.ts
@@ -63,6 +63,11 @@ router.delete('/:postId', authenticateJWT, async (req: Request, res: Response): 
       return;
     }
 
+    if (!postId){
+      res.status(401).json({ message: NOT_FOUND_POST.message});
+      return;
+    }
+
     const deletePostResponse = await postService.deletePost(userId, postId);
 
     if (deletePostResponse.isSuccess) {
